Fix Kotlin compilation errors in CommonComponents and ReportsScreen
- CommonComponents.kt:201 — score is Double, fillMaxWidth expects Float;
  fix: (score / 100.0).coerceIn(0.0, 1.0).toFloat()
- ReportsScreen.kt:62 — PaddingValues has no + operator; fix by wrapping
  LazyColumn in Box(Modifier.padding(paddingValues)) and using plain
  contentPadding = PaddingValues(16.dp)

+++ app/src/main/java/ir/alibahmani/ironorecalc/ui/components/CommonComponents.kt
@@ -198,7 +198,7 @@ fun ScoreBar(score: Double, modifier: Modifier = Modifier) {
         ) {
             Box(
                 modifier = Modifier
-                    .fillMaxWidth(fraction = (score / 100f).coerceIn(0f, 1f))
+                    .fillMaxWidth(fraction = (score / 100.0).coerceIn(0.0, 1.0).toFloat())
                     .fillMaxHeight()
                     .clip(RoundedCornerShape(4.dp))
                     .background(color)
